# PERF/WIP: CategoricalIndex.get_loc should avoid expensive cast to int64