--- a/pandas/_libs/algos.pyx
+++ b/pandas/_libs/algos.pyx
@@ -361,9 +361,13 @@ ctypedef fused algos_t:
     float64_t
     float32_t
     object
-    int32_t
     int64_t
+    int32_t
+    int16_t
+    int8_t
     uint64_t
+    uint32_t
+    uint16_t
     uint8_t