… doc/source/whatsnew/v0.24.0.txt