@WillAyd 2018-07-04
@topper-123 what issues were you running into with Cython? As far as generating the indexes goes I would think you could just add the appropriate types below:

https://github.com/pandas-dev/pandas/blob/2b136054db3b8f4c8d1aa33160c4bbe3b8ce940e/pandas/_libs/index_class_helper.pxi.in#L14

Not sure how you were planning to dynamically select the appropriate engine in `category.py` but as far as generating those engines in the Cython layer I think most of the foundation is already in place