@topper-123 2018-07-04
Ok, I think I understand the design now, at least. What is the templating language used and how is the pxi-file generated?

So would the correct approach be to:
* add int8_t, int16_t etc. to that list in index_helper_pxi.in,
* similarly add them to the list in algos.common_helper.pxi.in,
* ensure that CategoricalIndex._engine is of the correct type,
* avoid the  casting to int64 in  ``CategoricalIndex._engine``?