@topper-123 2018-07-04
Ok, got algos.common_helper.pxi.in compiling and works fine.

index_helper.pxi throws a compiler error. I added to the dtype list in index_helper.pxi.in:

```
          ('Int8', 'int8', 'int8_t'),
```

and get this error:

```cython
cythoning pandas\_libs/index.pyx to pandas\_libs\index.c

Error compiling Cython file:
------------------------------------------------------------
...

    cdef _maybe_get_bool_indexer(self, object val):
        cdef:
            ndarray[uint8_t, ndim=1, cast=True] indexer
            ndarray[intp_t, ndim=1] found
            ndarray[int8_t] values
                         ^
------------------------------------------------------------

pandas\_libs\index_class_helper.pxi:179:26: Invalid type.
````

It appears the compiler does not like the int8_t type. Using ``('UInt8', 'uint8', 'uint8_t'),`` instead compiles without errors.

Any ideas what's happening?