@jreback 2018-07-04
you need to do the imports of ``int8_t`` in ``index.pyx`` itself (see the top).