@jreback 2018-07-04
you might need to create the other hash tables  (definitions) as well
i think we only have 32 & 64 hits atm