@jreback 2018-07-05
hmm, i guess khash doesn't support this either.

ok, so maybe let's do a precursor PR to do this properly