@jreback 2018-07-08
@topper-123 well I suppose someone can do a PR to make other types of hashtables work. cc @pandas-dev/pandas-core