@topper-123 2018-07-08
That would be great. If someone makes the hash tables I will take this the rest of the way.