@topper-123 2018-07-31
I'v made a commit where I've solved the lack of Int8HashTable etc. by letting Int8Engine, Int16Engine and Int32Engine use Int64HashTable. All current tests pass and speed is good.

Is this an ok approach? I can then add some type specific tests if needed and update asv benchmarks.