@topper-123 2018-08-09
Hey, ``array.astype('int64', copy=False)`` won't work as explained above.

I don't see any good way forward other than someone commiting to make an Int8/16/32HashTable at some point. The question is thought if this PR needs to wait for that, or can be pulled in now and the relevant lines be changed when the hash table is implemented. 

I argue pulling this in now, as index operations are critical for performance and that would allow me to pursue a few more index-related performance issues.