@topper-123 2018-08-10
``algos.ensure_int64(values)`` ends up calling ``values.astype(np.int64)``, and therefore also causes the slowdown for int8/16/32 arrays that we want to avoid.

A *technically* possible solution to avoid that if statement while keeping the speed improvements would be to input a int64 array into _engine:

```python
    def _engine(self):
        codes = self.codes.astype('int64')
        return self._engine_type(lambda: codes, len(self))
```

But that would keep an int64 array in memory, in addition to an int8 array, so doesn't make any sense memorywise. We want to just use the existing int8 array.

I made another solution, where I've moved this specific implementation to index_class_helper.pxi.in. This makes its location more specific to the problem it solves.