@topper-123 2018-08-11
I've addressed the comments. Haven't made any new ASVs though, will look at it tomorrow, but  many of the new index engines are not currently used, so have to think about how to speed test them.