@topper-123 2018-08-11
The travis failure is the unexpected ResourceWarning, so unrelated to this PR.